set attribute0 colors in material test

+++ cts/header/ColorPalette.h
@@ -6,14 +6,14 @@
 namespace colors {
 extern const std::vector<glm::vec3> palette;
 
-static glm::vec3 getColorFromPalette(int index)
+static glm::vec3 getColorFromPalette(size_t index)
 {
   return palette.at(index % palette.size());
 }
 
-static std::vector<glm::vec3> getColorVectorFromPalette(int size) {
+static std::vector<glm::vec3> getColorVectorFromPalette(size_t size) {
   std::vector<glm::vec3> colors;
-  for (int i = 0; i < size; ++i) {
+  for (size_t i = 0; i < size; ++i) {
     colors.push_back(getColorFromPalette(i));
   }
   return colors;
